@@ -15,18 +15,18 @@
 		"typecheck": "tsc"
 	},
 	"dependencies": {
-		"@docusaurus/core": "^2.4.3",
-		"@docusaurus/preset-classic": "^2.4.3",
+		"@docusaurus/core": "3.0.0",
+		"@docusaurus/preset-classic": "3.0.0",
 		"@graphql-markdown/docusaurus": "^1.17.0",
 		"@graphql-tools/json-file-loader": "^7.4.17",
-		"@mdx-js/react": "^1.6.22",
+		"@mdx-js/react": "^3.0.0",
 		"clsx": "^1.2.1",
 		"docusaurus": "^1.14.7",
 		"graphql": "^16.6.0",
 		"prettier": "^2.8.7",
-		"prism-react-renderer": "^1.3.5",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2"
+		"prism-react-renderer": "^2.1.0",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0"
 	},
 	"browserslist": {
 		"production": [
@@ -41,12 +41,12 @@
 		]
 	},
 	"devDependencies": {
-		"@docusaurus/module-type-aliases": "^2.3.1",
-		"@tsconfig/docusaurus": "^1.0.5",
-		"typescript": "^4.7.4",
+		"@docusaurus/module-type-aliases": "3.0.0",
+		"@docusaurus/tsconfig": "3.0.0",
+		"typescript": "~5.2.2",
 		"yarn-audit-fix": "^7.2.3"
 	},
 	"engines": {
-		"node": ">=16.14"
+		"node": ">=18.0"
 	}
 }
